changed all virtual classes to virtualvariables.h
in order to acces meta information about all variables without needing to access hardware or data. it was made a change to the interfaces.

There is 3 variable class. One with meta information, one to deal with hardware, which requires meta, and another wich deals with data, that also requires meta.

A FAILED attempt was to make hardware and model classes inherit from meta. But that would still require to repeate code in each of the 3 interfaces.