Proper handling of exceptions.
Added deleting of Variables after catching an exception

Variables can trow in construction, in isReady and in read/set.
In some places, throw immediatly is not the proper way, so it tries 5 more times before throwing

Because produce() is in another thread, it catches internally, cleans and sends a message.